requirements.txt
django
django-filter
django-extensions
graphene-django
opentok
djangorestframework